+++ cmd/pint/watch.go
@@ -24,23 +24,17 @@ import (
 	"github.com/urfave/cli/v2"
 )
 
+const (
+	intervalFlag = "interval"
+	listenFlag   = "listen"
+	pidfileFlag  = "pidfile"
+)
+
 var watchCmd = &cli.Command{
 	Name:   "watch",
 	Usage:  "Continuously lint specified files",
 	Action: actionWatch,
 	Flags: []cli.Flag{
-		&cli.StringSliceFlag{
-			Name:    disabledFlag,
-			Aliases: []string{"d"},
-			Value:   cli.NewStringSlice(),
-			Usage:   "List of checks to disable (example: promql/cost)",
-		},
-		&cli.BoolFlag{
-			Name:    offlineFlag,
-			Aliases: []string{"o"},
-			Value:   false,
-			Usage:   "Disable all check that send live queries to Prometheus servers",
-		},
 		&cli.DurationFlag{
 			Name:    intervalFlag,
 			Aliases: []string{"i"},
@@ -52,6 +46,12 @@ var watchCmd = &cli.Command{
 			Aliases: []string{"s"},
 			Value:   ":8080",
 			Usage:   "Listen address for HTTP web server exposing metrics",
+		},
+		&cli.StringFlag{
+			Name:    pidfileFlag,
+			Aliases: []string{"p"},
+			Value:   "",
+			Usage:   "Write pid file to this path",
 		},
 	},
 }